Updated
Mostly working condition. I have error tested as much as I can, and
have fixed all errors that I have encountered. The only known “bug”
that has yet to be exterminated is that there is no cap to the depth of
the tree, which means you can create infinite levels. I will ask Mitra
if this is acceptable.